requirements.txt
adium-theme-ubuntu==0.3.4
alembic==0.9.8
apache-libcloud==2.3.0
APScheduler==3.4.0
asn1crypto==0.23.0
attrs==15.2.0
awscli==1.15.14
Babel==2.3.4
backports.csv==1.0.5
bcrypt==3.1.4
beautifulsoup4==4.4.1
blinker==1.3
boto3==1.7.14
botocore==1.10.14
cachetools==2.0.1
certifi==2018.4.16
cffi==1.11.2
chardet==3.0.4
cliapp==1.20160109
click==6.6
colorama==0.3.7
cryptography==2.1.4
cssselect==1.0.3
deluge==1.3.12
dill==0.2.8.2
Django==1.11
django-cors-headers==2.2.0
django-nose==1.4.5
django-webpack-loader==0.5.0
djangorestframework==3.7.7
docutils==0.14
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.21.0
extras==0.0.3
firebase-admin==2.9.1
fixtures==2.0.0
Flask==0.11.1
Flask-Babel==0.11.1
Flask-Gravatar==0.4.2
Flask-HTMLmin==1.3.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.0.3
Flask-Paranoid==0.1.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==1.7.5
Flask-SocketIO==2.9.3
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
funcsigs==1.0.2
future==0.16.0
futures==3.2.0
gax-google-logging-v2==0.8.3
gax-google-pubsub-v1==0.8.3
gcloud==0.17.0
geographiclib==1.49
geopy==1.16.0
google-api-core==1.1.0
google-auth==1.4.1
google-cloud-core==0.28.1
google-cloud-firestore==0.29.0
google-cloud-storage==1.8.0
google-gax==0.16.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
greenlet==0.4.12
grpc-google-logging-v2==0.8.1
grpc-google-pubsub-v1==0.8.1
grpcio==1.14.2
gunicorn==19.7.1
gyp==0.1
html5lib==1.0b3
htmlmin==0.1.12
httplib2==0.11.3
humanfriendly==4.10
idna==2.6
importlib==1.0.4
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.7.3
jmespath==0.9.3
jsonpickle==0.9.6
jws==0.1.3
linecache2==1.0.0
lml==0.0.1
lxml==4.2.0
Mako==1.0.7
Markdown==2.6.6
MarkupSafe==0.23
monotonic==1.4
nltk==3.2.5
nose==1.3.7
numpy==1.11.0
oauth2client==3.0.0
odfpy==1.3.6
PAM==0.4.2
pandas==0.22.0
paramiko==2.4.0
parsel==1.4.0
passlib==1.6.2
pbr==1.9.1
pexpect==4.3.0
pgadmin4==2.1
Pillow==5.0.0
ply==3.8
protobuf==3.5.2
psycopg2==2.7.3.2
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.4.3
PyDispatcher==2.0.5
pyexcel-io==0.5.6
pyexcel-ods==0.5.2
pygame===1.9.1release
Pygments==2.1
pygobject==3.20.0
PyNaCl==1.2.1
pyOpenSSL==16.2.0
Pyrebase==3.0.27
pyrsistent==0.11.13
pyserial==3.0.1
python-dateutil==2.5.0
python-digitalocean==1.13.2
python-editor==1.0.3
python-engineio==2.0.1
python-jwt==2.0.1
python-mimeparse==1.5.1
python-socketio==1.8.4
python2==1.2
pytz==2014.10
pyxdg==0.25
PyYAML==3.11
queuelib==1.5.0
requests==2.11.1
requests-toolbelt==0.7.0
rsa==3.4.2
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.0.0
Scrapy==1.5.0
service-identity==16.0.0
simplejson==3.6.5
six==1.11.0
sklearn==0.0
speaklater==1.3
SQLAlchemy==1.0.14
sqlparse==0.1.19
testtools==2.3.0
traceback2==1.4.0
ttystatus==0.32
Twisted==16.0.0
tzlocal==1.5.1
Unidecode==1.0.22
unittest2==1.1.0
unity-lens-photos==1.0
urllib3==1.22
vboxapi==1.0
w3lib==1.19.0
Werkzeug==0.9.6
WTForms==2.0.2
xlrd==1.1.0
zope.interface==4.1.3
